# Allow to obtain informations of used direct and heap memory for ByteBufAllocator implementations