Motivation:

Often its useful for the user to be able to get some stats about the memory allocated via an allocator.

Modifications:

Allow to obtain the used heap and direct memory for an allocator
Add test case

Result:

Fixes [#6341]