@CodingFabian 2017-02-27
cannot judge about the code change, it looks a bit "long" for such a "simple" task.

The new API methods would be really helpful for us. Thanks