@normanmaurer 2017-02-27
@CodingFabian what you mean with "the new API methods" ? The ones that I added to `AbstractByteBufAllocator` ?