@trustin 2017-02-28
How about just adding two new methods to `ByteBufAllocator` and implement them in all implementations? If API compatibility is a concern, we could define a new interface that extends `ByteBufAllocator` and add them there and make all our implementations implement it in lieu of `ByteBufAllocator`.