@normanmaurer 2017-02-28
Yeah that is the concern... do you have a good interface name @trustin?

> Am 28.02.2017 um 10:30 schrieb Trustin Lee <notifications@github.com>:
> 
> How about just adding two new methods to ByteBufAllocator and implement them in all implementations? If API compatibility is a concern, we could define a new interface that extends ByteBufAllocator and add them there and make all our implementations implement it in lieu of ByteBufAllocator.
> 
> —
> You are receiving this because you were assigned.
> Reply to this email directly, view it on GitHub, or mute the thread.
>