@trustin 2017-02-28
> Yeah that is the concern... do you have a good interface name @trustin?

`ByteBufAllocatorStats`? and add a bunch of stats there with some default values that indicates 'unsupported'