@normanmaurer 2017-02-28
@trustin maybe InstrumentedByteBufAllocator ?

> Am 28.02.2017 um 12:04 schrieb Trustin Lee <notifications@github.com>:
> 
> Yeah that is the concern... do you have a good interface name @trustin?
> 
> ByteBufAllocatorStats? and add a bunch of stats there with some default values that indicates 'unsupported'
> 
> —
> You are receiving this because you were assigned.
> Reply to this email directly, view it on GitHub, or mute the thread.
>