@normanmaurer 2017-02-28
@trustin actually let me call it `ByteBufAllocatorMetric` to be consistent with the other Metric interfaces that are exposed by the `PooledByteBufAllocator`.