@normanmaurer 2017-03-01
@Scottmitch sure I can... that said when I did last time I was not able to see any perf impact.