@normanmaurer 2017-03-01
@Scottmitch I did a benchmark with only the `UnpooledByteBufAllocator` as this is the only one which changed and only allocate direct buffers as if allocating heap buffers in this manner it will be dominated by GC anyway.

Here are the results:

Before the PR:
```
# Run complete. Total time: 00:02:07

Benchmark                                            (size)   Mode  Cnt        Score        Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateRelease   00064  thrpt   20  6679802.479 ± 248202.037  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   00256  thrpt   20  6766225.952 ± 278366.008  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   01024  thrpt   20  6464984.329 ± 117666.813  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   04096  thrpt   20  6592483.889 ± 136106.843  ops/s

# Run complete. Total time: 00:02:07

Benchmark                                            (size)   Mode  Cnt        Score        Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateRelease   00064  thrpt   20  6560584.865 ± 103409.666  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   00256  thrpt   20  7177342.418 ± 317870.561  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   01024  thrpt   20  6568157.769 ±  88297.714  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   04096  thrpt   20  6612075.957 ± 105282.223  ops/s
```

With the PR:
```
# Run complete. Total time: 00:02:07

Benchmark                                            (size)   Mode  Cnt        Score        Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateRelease   00064  thrpt   20  6498802.876 ±  52565.274  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   00256  thrpt   20  6683246.671 ±  55994.734  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   01024  thrpt   20  6841686.396 ± 278991.411  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   04096  thrpt   20  6523478.292 ±  27012.837  ops/s

# Run complete. Total time: 00:02:07

Benchmark                                            (size)   Mode  Cnt        Score        Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateRelease   00064  thrpt   20  6622965.147 ± 249237.717  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   00256  thrpt   20  6848344.387 ± 294846.614  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   01024  thrpt   20  6809770.757 ± 157764.356  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateRelease   04096  thrpt   20  6553908.127 ± 127681.444  ops/s
```