@normanmaurer 2017-03-01
@Scottmitch ... I will squash the benchmark into the commit as well before pull in