@normanmaurer 2017-03-01
Yep

> Am 01.03.2017 um 15:42 schrieb Johno Crawford <notifications@github.com>:
> 
> @johnou commented on this pull request.
> 
> In buffer/src/main/java/io/netty/buffer/InstrumentedByteBufAllocator.java:
> 
> > + * with the License. You may obtain a copy of the License at:
> + *
> + *   http://www.apache.org/licenses/LICENSE-2.0
> + *
> + * Unless required by applicable law or agreed to in writing, software
> + * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
> + * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
> + * License for the specific language governing permissions and limitations
> + * under the License.
> + */
> +package io.netty.buffer;
> +
> +/**
> + * {@link ByteBufAllocator} which exposes metrics.
> + */
> +public interface InstrumentedByteBufAllocator extends ByteBufAllocator {
> @normanmaurer did you change your mind about calling it ByteBufAllocatorMetric?
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.
>