@normanmaurer 2017-03-01
Cherry-picked into 4.1 (461f9a1) and 4.0 (57fd316)